[* As Code] move meta schema to @kbn/as-code-shared-schemas package
