+++ src/platform/packages/shared/as-code/shared-schemas/index.ts
@@ -7,4 +7,4 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export { asCodeIdSchema } from './src/schemas';
+export { asCodeIdSchema, asCodeMetaSchema, getMeta, asCodeResolveSchema } from './src/schemas';
