+++ src/platform/packages/shared/as-code/shared-schemas/src/schemas/index.ts
@@ -8,3 +8,5 @@
  */
 
 export { asCodeIdSchema } from './id';
+export { asCodeMetaSchema, getMeta } from './meta';
+export { asCodeResolveSchema } from './resolve_schema';
