+++ src/platform/packages/shared/as-code/shared-schemas/src/schemas/resolve_schema.ts
@@ -8,13 +8,22 @@
  */
 
 import { schema } from '@kbn/config-schema';
-import { asCodeMetaSchema } from '@kbn/as-code-shared-schemas';
-import { markdownAttributesSchema } from '../../markdown_saved_object/schema/v1';
 
-export const createRequestBodySchema = markdownAttributesSchema;
-
-export const createResponseBodySchema = schema.object({
-  id: schema.string(),
-  data: markdownAttributesSchema,
-  meta: asCodeMetaSchema,
-});
+export const asCodeResolveSchema = schema.object(
+  {
+    outcome: schema.oneOf([
+      schema.literal('exactMatch'),
+      schema.literal('aliasMatch'),
+      schema.literal('conflict'),
+    ]),
+    alias_target_id: schema.maybe(schema.string()),
+    alias_purpose: schema.maybe(
+      schema.oneOf([schema.literal('savedObjectConversion'), schema.literal('savedObjectImport')])
+    ),
+  },
+  {
+    meta: {
+      id: 'kbn-as-code-resolve',
+    },
+  }
+);
