+++ src/platform/packages/shared/as-code/shared-schemas/tsconfig.json
@@ -11,5 +11,6 @@
   ],
   "kbn_references": [
     "@kbn/config-schema",
+    "@kbn/core-saved-objects-common",
   ]
 }
