+++ src/platform/plugins/shared/dashboard/server/api/create/schemas.ts
@@ -8,8 +8,8 @@
  */
 
 import { schema } from '@kbn/config-schema';
+import { asCodeMetaSchema } from '@kbn/as-code-shared-schemas';
 import { getDashboardStateSchema } from '../dashboard_state_schemas';
-import { baseMetaSchema, createdMetaSchema, updatedMetaSchema } from '../meta_schemas';
 
 export function getCreateRequestBodySchema(isDashboardAppRequest: boolean) {
   return getDashboardStateSchema(isDashboardAppRequest);
@@ -19,6 +19,6 @@ export function getCreateResponseBodySchema(isDashboardAppRequest: boolean) {
   return schema.object({
     id: schema.string(),
     data: getDashboardStateSchema(isDashboardAppRequest),
-    meta: schema.allOf([baseMetaSchema, createdMetaSchema, updatedMetaSchema]),
+    meta: asCodeMetaSchema,
   });
 }
