+++ src/platform/plugins/shared/dashboard_markdown/server/api/read/schemas.ts
@@ -8,11 +8,11 @@
  */
 
 import { schema } from '@kbn/config-schema';
-import { baseMetaSchema, createdMetaSchema, updatedMetaSchema } from '../meta_schemas';
+import { asCodeMetaSchema } from '@kbn/as-code-shared-schemas';
 import { markdownAttributesSchema } from '../../markdown_saved_object/schema/v1';
 
 export const readResponseBodySchema = schema.object({
   id: schema.string(),
   data: markdownAttributesSchema,
-  meta: schema.allOf([baseMetaSchema, createdMetaSchema, updatedMetaSchema]),
+  meta: asCodeMetaSchema,
 });
